setup/requirements.txt
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
decorator==4.4.2  # This specific version worked with moviepy
idna==3.10
imageio>=2.31.1
imageio-ffmpeg>=0.4.8
joblib==1.4.2
lazy_loader==0.4
librosa>=0.10.2
llvmlite==0.44.0
moviepy==2.0.0.dev2
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
opencv-python>=4.11.0
packaging==24.2
Pillow>=10.0.0
platformdirs==4.3.6
pooch==1.8.2
proglog<=1.0.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
soundfile>=0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0

